--- a/lib/benchmark.js
+++ b/lib/benchmark.js
@@ -1,6 +1,6 @@
 
 const colors = require( './string-colors' )
-const { Lines, Banis } = require( '../' )
+const { Lines, Banis, Shabads } = require( '../' )
 
 const BENCHMARK_COUNT = 10
 const MS_THRESHOLD = 50
@@ -13,15 +13,29 @@ const fullWordSearch = search => Lines.query().fullWord( search )
 
 const getBanisWithSections = () => Banis.query()
 
-const getBaniLines = baniId => Banis
+const getBaniLines = async baniId => Banis
   .query()
-  .joinEager( 'lines.shabad' )
-  .orderBy( [ 'line_group', 'l.order_id' ] )
+  .joinEager( 'lines' )
+  .orderBy( [ 'line_group', 'lines.order_id' ] )
   .where( 'banis.id', baniId )
   .withTranslations()
   .withTransliterations()
-  .eagerOptions( { minimize: false, aliases: { lines: 'l' } } )
   .then( ( [ bani ] ) => bani )
+
+const getShabad = async shabadId => Shabads
+  .query()
+  .joinEager( 'lines' )
+  .orderBy( 'lines.order_id' )
+  .where( 'shabads.id', shabadId )
+  .withTranslations()
+  .withTransliterations()
+  .then( ( [ shabad ] ) => shabad )
+
+const getLine = async lineId => Lines
+  .query()
+  .where( 'lines.id', lineId )
+  .withTranslations()
+  .then( ( [ line ] ) => line )
 
 // Runs an async function count times and returns the average
 const benchmark = async ( fn, count = BENCHMARK_COUNT ) => {
@@ -58,6 +72,20 @@ const main = async () => {
   for ( const search of fullWordSearches ) {
     const { average, res } = await benchmark( () => fullWordSearch( search ) )
     console.log( `${search}:`.value, msColor( average ), `(${res.length} lines)` )
+  }
+
+  const lineIds = [ 'KNCN', 'DXUY', 'PZ9Q', 'LZPR', 'QD98', 'XQ82' ]
+  console.log( '\nLines'.subheader )
+  for ( const lineId of lineIds ) {
+    const { average } = await benchmark( () => getLine( lineId ) )
+    console.log( `${lineId}:`.value, msColor( average ) )
+  }
+
+  const shabadIds = [ '9AC', 'N1W', 'GE6', 'B6R', 'HHZ', '0T2' ]
+  console.log( '\nShabads'.subheader )
+  for ( const shabadId of shabadIds ) {
+    const { average, res } = await benchmark( () => getShabad( shabadId ) )
+    console.log( `${shabadId}:`.value, msColor( average ), `(${res.lines.length} lines)` )
   }
 
   console.log( '\nAll Banis'.subheader )