--- a/lib/benchmark.js
+++ b/lib/benchmark.js
@@ -35,6 +35,7 @@ const getLine = async lineId => Lines
   .query()
   .where( 'lines.id', lineId )
   .withTranslations()
+  .withTransliterations()
   .then( ( [ line ] ) => line )
 
 // Runs an async function count times and returns the average