--- a/lib/models/Banis.js
+++ b/lib/models/Banis.js
@@ -26,6 +26,7 @@ class Banis extends BaseModel {
           through: {
             from: 'bani_lines.bani_id',
             to: 'bani_lines.line_id',
+            extra: [ 'line_group' ],
           },
           to: 'lines.id',
         },