--- a/lib/models/Banis.js
+++ b/lib/models/Banis.js
@@ -8,13 +8,31 @@ const { Model } = require( 'objection' )
 const BaseModel = require( './BaseModel' )
 const CommonQueryBuilderFactory = require( './CommonQueryBuilder' )
 
+class BaniQueryBuilder extends CommonQueryBuilderFactory( 'lines' ) {
+  static getLineIds( bani ) {
+    return Model
+      .knex()
+      .select( 'line_id' )
+      .from( 'bani_lines' )
+      .where( 'bani_id', bani.id )
+  }
+
+  withTranslations() {
+    return super.appendTranslations( BaniQueryBuilder.getLineIds )
+  }
+
+  withTransliterations() {
+    return super.appendTransliterations( BaniQueryBuilder.getLineIds )
+  }
+}
+
 class Banis extends BaseModel {
   static get tableName() {
     return 'banis'
   }
 
   static get QueryBuilder() {
-    return CommonQueryBuilderFactory( 'lines' )
+    return BaniQueryBuilder
   }
 
   static get relationMappings() {