--- a/lib/models/Lines.js
+++ b/lib/models/Lines.js
@@ -13,6 +13,18 @@ const BaseModel = require( './BaseModel' )
 const CommonQueryBuilder = require( './CommonQueryBuilder' )()
 
 class LineQueryBuilder extends CommonQueryBuilder {
+  static getLineIds( line ) {
+    return [ line.id ]
+  }
+
+  withTranslations() {
+    return super.appendTranslations( LineQueryBuilder.getLineIds )
+  }
+
+  withTransliterations() {
+    return super.appendTransliterations( LineQueryBuilder.getLineIds )
+  }
+
   /**
    * Given a search query and column name, return a query that searches for the shabad.
    * @param search The query text to search for a partial match of.