--- a/lib/models/Shabads.js
+++ b/lib/models/Shabads.js
@@ -7,6 +7,24 @@ const { Model } = require( 'objection' )
 
 const BaseModel = require( './BaseModel' )
 const CommonQueryBuilderFactory = require( './CommonQueryBuilder' )
+const Lines = require( './Lines' )
+
+class ShabadQueryBuilder extends CommonQueryBuilderFactory( 'lines' ) {
+  static getLineIds( shabad ) {
+    return Lines
+      .query()
+      .select( 'id' )
+      .where( 'shabad_id', shabad.id )
+  }
+
+  withTranslations() {
+    return super.appendTranslations( ShabadQueryBuilder.getLineIds )
+  }
+
+  withTransliterations() {
+    return super.appendTransliterations( ShabadQueryBuilder.getLineIds )
+  }
+}
 
 class Shabads extends BaseModel {
   static get tableName() {
@@ -14,7 +32,7 @@ class Shabads extends BaseModel {
   }
 
   static get QueryBuilder() {
-    return CommonQueryBuilderFactory( 'lines' )
+    return ShabadQueryBuilder
   }
 
   static get relationMappings() {