Without translations and transliterations:
```
Running benchmarks
10 samples will be collected

First Letter Searching
hh: 77.6ms (3115 lines)
hhh: 30.5ms (253 lines)
hhhh: 25.9ms (49 lines)
hhhgg: 48.4ms (6 lines)
h_r_p: 55.7ms (40 lines)

Full Word Searching
ihr: 53.1ms (753 lines)
jo jnu: 32ms (21 lines)

Lines
KNCN: 3.7ms
DXUY: 1.9ms
PZ9Q: 2.3ms
LZPR: 2.7ms
QD98: 2ms
XQ82: 1.6ms

Shabads
9AC: 2.6ms (19 lines)
N1W: 2.5ms (15 lines)
GE6: 2.3ms (8 lines)
B6R: 2.1ms (23 lines)
HHZ: 1.6ms (13 lines)
0T2: 1.6ms (7 lines)

All Banis
All Banis: 1.4ms (25 results)

All Bani Lines
Jap Ji Sahib: 15.5ms (385 lines)
Jaap Sahib: 26.4ms (801 lines)
Tav Prasad Svaiyay: 3.5ms (43 lines)
Benti Choupai Sahib: 5.7ms (104 lines)
Anand Sahib: 6.7ms (210 lines)
Anand Sahib (6 Pauris and Salok): 4ms (41 lines)
Rehraas Sahib: 10.5ms (340 lines)
Aartee: 4.1ms (84 lines)
Sohila Sahib: 3.5ms (54 lines)
Sukhmani Sahib: 57.5ms (2027 lines)
Assa Ki Vaar: 20.7ms (734 lines)
Lava(n): 2.6ms (25 lines)
Akhand Path Bhog (Mundavani): 4.7ms (128 lines)
Akhand Path Bhog (Raagmala): 6.5ms (190 lines)
Ardaas: 2.2ms (27 lines)
Shabad Hazaare: 4.7ms (103 lines)
Shabad Hazaare Patishahi 10: 4.2ms (87 lines)
Tav Prasad Svaiyay (Deenan Kee): 2.8ms (43 lines)
Akal Ustat: 32.6ms (1101 lines)
Bavan Akhree: 20.6ms (689 lines)
Sidh Gosht: 12.9ms (406 lines)
Oankaar: 12.1ms (399 lines)
Barah Maha: 4.9ms (129 lines)
Salok Mehla 9: 5.1ms (117 lines)
Chandi Di Vaar: 11.2ms (389 lines)
```

With original translations and transliterations:
```
Running benchmarks
10 samples will be collected

First Letter Searching
hh: 64.8ms (3115 lines)
hhh: 28.6ms (253 lines)
hhhh: 25.1ms (49 lines)
hhhgg: 25.4ms (6 lines)
h_r_p: 26.6ms (40 lines)

Full Word Searching
ihr: 44.1ms (753 lines)
jo jnu: 30.3ms (21 lines)

Lines
KNCN: 3.9ms
DXUY: 2.4ms
PZ9Q: 1.8ms
LZPR: 2.1ms
QD98: 2ms
XQ82: 2.1ms

Shabads
9AC: 22.3ms (19 lines)
N1W: 17.1ms (15 lines)
GE6: 4.4ms (8 lines)
B6R: 23.9ms (23 lines)
HHZ: 13.7ms (13 lines)
0T2: 8.2ms (7 lines)

All Banis
All Banis: 1.2ms (25 results)

All Bani Lines
Jap Ji Sahib: 381.2ms (385 lines)
Jaap Sahib: 241.6ms (801 lines)
Tav Prasad Svaiyay: 16.9ms (43 lines)
Benti Choupai Sahib: 31.5ms (104 lines)
Anand Sahib: 190.5ms (210 lines)
Anand Sahib (6 Pauris and Salok): 36.3ms (41 lines)
Rehraas Sahib: 243.3ms (340 lines)
Aartee: 57.5ms (84 lines)
Sohila Sahib: 47.7ms (54 lines)
Sukhmani Sahib: 1880.8ms (2027 lines)
Assa Ki Vaar: 666.9ms (734 lines)
Lava(n): 25.1ms (25 lines)
Akhand Path Bhog (Mundavani): 123.8ms (128 lines)
Akhand Path Bhog (Raagmala): 173.5ms (190 lines)
Ardaas: 8.9ms (27 lines)
Shabad Hazaare: 88.5ms (103 lines)
Shabad Hazaare Patishahi 10: 29.2ms (87 lines)
Tav Prasad Svaiyay (Deenan Kee): 15.7ms (43 lines)
Akal Ustat: 356.6ms (1101 lines)
Bavan Akhree: 621.1ms (689 lines)
Sidh Gosht: 356.6ms (406 lines)
Oankaar: 361.2ms (399 lines)
Barah Maha: 121.7ms (129 lines)
Salok Mehla 9: 106.7ms (117 lines)
Chandi Di Vaar: 124.5ms (389 lines)
```

With new translations:
```
Running benchmarks
10 samples will be collected

First Letter Searching
hh: 62ms (3115 lines)
hhh: 27.5ms (253 lines)
hhhh: 26.4ms (49 lines)
hhhgg: 29.7ms (6 lines)
h_r_p: 24.4ms (40 lines)

Full Word Searching
ihr: 42.9ms (753 lines)
jo jnu: 29.5ms (21 lines)

Lines
KNCN: 1.8ms
DXUY: 1.6ms
PZ9Q: 1.6ms
LZPR: 1.6ms
QD98: 1.6ms
XQ82: 1.3ms

Shabads
9AC: 5.1ms (19 lines)
N1W: 3.3ms (15 lines)
GE6: 2.8ms (8 lines)
B6R: 4.6ms (23 lines)
HHZ: 3.2ms (13 lines)
0T2: 2.4ms (7 lines)

All Banis
All Banis: 1.3ms (25 results)

All Bani Lines
Jap Ji Sahib: 32.7ms (385 lines)
Jaap Sahib: 33.6ms (801 lines)
Tav Prasad Svaiyay: 5.7ms (43 lines)
Benti Choupai Sahib: 8.7ms (104 lines)
Anand Sahib: 17.4ms (210 lines)
Anand Sahib (6 Pauris and Salok): 5.4ms (41 lines)
Rehraas Sahib: 22.6ms (340 lines)
Aartee: 7.1ms (84 lines)
Sohila Sahib: 6.8ms (54 lines)
Sukhmani Sahib: 131.5ms (2027 lines)
Assa Ki Vaar: 48.8ms (734 lines)
Lava(n): 3.5ms (25 lines)
Akhand Path Bhog (Mundavani): 9ms (128 lines)
Akhand Path Bhog (Raagmala): 11.1ms (190 lines)
Ardaas: 2.6ms (27 lines)
Shabad Hazaare: 7.8ms (103 lines)
Shabad Hazaare Patishahi 10: 4.5ms (87 lines)
Tav Prasad Svaiyay (Deenan Kee): 3.6ms (43 lines)
Akal Ustat: 43.7ms (1101 lines)
Bavan Akhree: 47.7ms (689 lines)
Sidh Gosht: 29.2ms (406 lines)
Oankaar: 31.4ms (399 lines)
Barah Maha: 9.4ms (129 lines)
Salok Mehla 9: 8.3ms (117 lines)
Chandi Di Vaar: 17.6ms (389 lines)
```

With new translations and transliterations:
```
Running benchmarks
10 samples will be collected

First Letter Searching
hh: 65.9ms (3115 lines)
hhh: 28.1ms (253 lines)
hhhh: 24.9ms (49 lines)
hhhgg: 24.9ms (6 lines)
h_r_p: 30.7ms (40 lines)

Full Word Searching
ihr: 67.8ms (753 lines)
jo jnu: 33.6ms (21 lines)

Lines
KNCN: 3.5ms
DXUY: 2.4ms
PZ9Q: 2.6ms
LZPR: 2.1ms
QD98: 3.2ms
XQ82: 2.7ms

Shabads
9AC: 7ms (19 lines)
N1W: 4.8ms (15 lines)
GE6: 4.2ms (8 lines)
B6R: 5ms (23 lines)
HHZ: 3.6ms (13 lines)
0T2: 3.3ms (7 lines)

All Banis
All Banis: 0.8ms (25 results)

All Bani Lines
Jap Ji Sahib: 40.5ms (385 lines)
Jaap Sahib: 53.3ms (801 lines)
Tav Prasad Svaiyay: 5.1ms (43 lines)
Benti Choupai Sahib: 8.5ms (104 lines)
Anand Sahib: 23.4ms (210 lines)
Anand Sahib (6 Pauris and Salok): 6.6ms (41 lines)
Rehraas Sahib: 32.4ms (340 lines)
Aartee: 9.5ms (84 lines)
Sohila Sahib: 7.9ms (54 lines)
Sukhmani Sahib: 182.3ms (2027 lines)
Assa Ki Vaar: 65.9ms (734 lines)
Lava(n): 5.7ms (25 lines)
Akhand Path Bhog (Mundavani): 14.6ms (128 lines)
Akhand Path Bhog (Raagmala): 24.6ms (190 lines)
Ardaas: 5.2ms (27 lines)
Shabad Hazaare: 38.9ms (103 lines)
Shabad Hazaare Patishahi 10: 10.8ms (87 lines)
Tav Prasad Svaiyay (Deenan Kee): 8.2ms (43 lines)
Akal Ustat: 70.6ms (1101 lines)
Bavan Akhree: 64.9ms (689 lines)
Sidh Gosht: 38.1ms (406 lines)
Oankaar: 39.4ms (399 lines)
Barah Maha: 13.2ms (129 lines)
Salok Mehla 9: 10.8ms (117 lines)
Chandi Di Vaar: 26.6ms (389 lines)
```